pyproject.toml
[tool.poetry]
name = "office-connect"
version = "0.1.0"
description = "Office 365 / MS Graph logic layer with MCP server support"
authors = ["Michael Ikemann"]
license = "MIT"
packages = [{include = "office_con"}]

[tool.poetry.dependencies]
python = "^3.12"
O365 = ">2.0.26"
pydantic = ">2.5.3"
python-dotenv = ">1.0.0"
redis = "*"
pymongo = "*"
beautifulsoup4 = ">4.12.3"
pymupdf = "*"
PyJWT = "*"
aiohttp = "*"
cryptography = "*"
bcrypt = "*"
requests = "*"
mcp = ">=1.0.0"
fastapi = "^0.135.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.5"
mypy = "^1.19.1"
pytest-cov = "^7.0.0"
pytest-asyncio = "^1.3.0"
uvicorn = {version = ">=0.30", extras = ["standard"]}

[tool.poetry.scripts]
office-connect = "office_con.mcp_server:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
